cleanup error pages
