+++ Sources/Vapor/Error/ErrorRenderer.swift
@@ -1,0 +1,5 @@
+import HTTP
+
+public protocol ErrorRenderer {
+    func make(with req: Request, for error: Error) -> Response
+}
