+++ Tests/VaporTests/ErrorTests.swift
@@ -1,5 +1,6 @@
 import XCTest
 @testable import Vapor
+import HTTP
 
 class ErrorTests: XCTestCase {
     static let allTests = [
@@ -8,8 +9,9 @@ class ErrorTests: XCTestCase {
 
     func testFixes() throws {
         let drop = try Droplet()
-        let view = drop.view.make(Abort(.notFound))
-        XCTAssert(view.data.makeString().contains("404"))
-        XCTAssert(view.data.makeString().contains("Not Found"))
+        let req = try Request(method: .get, uri: "foo", headers: ["Accept": "html"])
+        let view = drop.errorRenderer.make(with: req, for: Abort(.notFound))
+        XCTAssert(try view.bodyString().contains("404"))
+        XCTAssert(try view.bodyString().contains("Not Found"))
     }
 }
