+++ Tests/VaporTests/Utilities/DebugConsole.swift
@@ -16,6 +16,12 @@ class DebugConsole: ConsoleProtocol {
         inputBuffer = ""
         return temp
     }
+    
+    
+    func secureInput() -> String {
+        return input()
+    }
+
 
     func output(_ string: String, style: ConsoleStyle, newLine: Bool) {
         outputBuffer += string
